[Base][Refactoring] Config validation and decompose main.py
1. Fix HF_HOME crash when env var is unset
2. Validate required cluster variables after loading clusters.yaml
3. Replace hardcoded lmms-eval adapter detection with LMMS_MODEL_ADAPTERS lookup
4. Remove magic-number time budgeting to use TIME_LIMIT from clusters.yaml directly
5. Decompose main.py into constants.py and results.py
